feat(nixify): use nixify without creating new instance
