+++ @Nixify/core/nixify.d.ts
@@ -1,0 +1,7 @@
+import { ServiceConfig, XOR, ServiceReqMethods, RequestMethods } from '../types/index.js';
+
+declare const _default: {
+    create: <T extends ServiceConfig>(config?: T) => XOR<ServiceReqMethods<T>, RequestMethods>;
+};
+
+export { _default as default };
