+++ @Nixify/index.d.ts
@@ -1,7 +1,7 @@
-import { ServiceConfig, XOR, ServiceReqMethods, RequestMethods } from './types/index.js';
+import { RequestMethods, ServiceConfig, XOR, ServiceReqMethods } from './types/index.js';
 
-declare const _default: {
+declare const _default: (RequestMethods & Record<string, never>) | ({
     create: <T extends ServiceConfig>(config?: T) => XOR<ServiceReqMethods<T>, RequestMethods>;
-};
+} & Record<string, never>);
 
 export { _default as default };
