+++ @Nixify/utils/index.d.ts
@@ -15,8 +15,6 @@ declare const mergeConfigs: (config: Options, methodConfig: MethodConfig, method
     abortController: AbortController;
     path?: string;
     qs?: {
-        [name: string]: number | queryType;
-    } | {
         readonly strict?: boolean;
         readonly encode?: boolean;
         readonly arrayFormat?: "bracket" | "index" | "comma" | "separator" | "bracket-separator" | "colon-list-separator" | "none";
@@ -24,6 +22,8 @@ declare const mergeConfigs: (config: Options, methodConfig: MethodConfig, method
         readonly sort?: false | ((itemLeft: string, itemRight: string) => number);
         readonly skipNull?: boolean;
         readonly skipEmptyString?: boolean;
+    } | {
+        [name: string]: number | queryType;
     };
     json?: object;
     responseType?: ResponseType;
